backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Backend for Todo App with FastAPI and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
sqlmodel = "^0.0.8"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pyjwt = "^2.8.0"
pydantic = "^2.5.0"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
neon = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"